backend/requirements.txt
# FastAPI and server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.25
asyncpg>=0.29.0
alembic>=1.13.1
psycopg2-binary>=2.9.9

# Redis
redis>=5.0.1

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.1.2

# Validation
pydantic>=2.5.3
pydantic-settings>=2.1.0
email-validator>=2.1.0

# File handling
python-magic>=0.4.27
aiofiles>=23.2.1
Pillow>=10.2.0
pdf2image>=1.17.0

# PDF generation
reportlab>=4.0.8
weasyprint>=60.2

# Excel/CSV
openpyxl>=3.1.2
pandas>=2.1.4

# AI
anthropic>=0.18.0

# Email
aiosmtplib>=3.0.1

# Utilities
python-dateutil>=2.8.2
pytz>=2024.1

# Testing
pytest>=7.4.4
pytest-asyncio>=0.23.3
httpx>=0.26.0

# Development
black>=24.1.1
isort>=5.13.2
flake8>=7.0.0
